+++ src/mediaserver_information.rs
@@ -1,3 +1,4 @@
+#![allow(non_snake_case)]
 extern crate getch;
 use std::fmt::Debug;
 use std::io;
@@ -11,6 +12,7 @@ use isahc::Request;
 use isahc::prelude::*;
 use serde_json::Value;
 use std::time::Duration;
+use rpassword::read_password;
 use std::io::prelude::*;
 use std::net::UdpSocket;
 use std::str::from_utf8;
@@ -18,9 +20,9 @@ use crate::APPNAME;
 use crate::VERSION;
 use crate::APP_INFO;
 use crate::settings::Settings;
+use crate::config::*;
 use std::result::Result;
 use serde_derive::{Deserialize,Serialize};
-use std::path::Path;
 
 
 #[derive(Debug)]
@@ -34,11 +36,28 @@ pub struct HeadDict {
 }
 
 
-#[derive(Debug, Deserialize, Serialize)]
+#[derive(Debug, Deserialize, Serialize, Clone)]
 pub struct ConfigFile {
     pub emby: bool,
     pub ipaddress: String,
     pub device_id: String,
+    pub user_id: String,
+    pub access_token: String,
+    pub username: String
+}
+
+
+#[derive(Debug, Deserialize, Serialize, Clone)]
+pub struct ConfigFileRaw {
+    pub emby: bool,
+    pub ipaddress: String,
+    pub device_id: String,
+    pub user: Vec<ConfigFileUser>
+}
+
+
+#[derive(Debug, Deserialize, Serialize, Clone)]
+pub struct ConfigFileUser {
     pub user_id: String,
     pub access_token: String,
     pub username: String
@@ -94,17 +113,17 @@ pub fn check_information(settings: &Settings) -> HeadDict {
     let mut auth_header: AuthHeader;
     let device_id = uuid::Uuid::new_v4().to_string();
     let server_kind = if settings.server_config.is_none() {
-        print!("What kind of server do you want to stream from?\n [1] Emby\n [2] Jellyfin");
+        print!("What kind of server do you want to stream from?\n   [1] Emby\n   [2] Jellyfin");
         getch("12")
     } else {
-        let config_file = read_config(settings.server_config.as_ref().unwrap(), true);
-        if config_file.is_some() {
-            match config_file.unwrap().emby {
+        let config_file = read_config(settings.server_config.as_ref().unwrap(), settings.autologin);
+        if config_file.is_ok() {
+            match config_file.unwrap().0.emby {
                 true => '1',
                 false => '2'
             }
         } else {
-            print!("What kind of server do you want to stream from?\n [1] Emby\n [2] Jellyfin");
+            print!("What kind of server do you want to stream from?\n   [1] Emby\n   [2] Jellyfin");
             getch("12")
         }
     };
@@ -126,23 +145,23 @@ pub fn check_information(settings: &Settings) -> HeadDict {
             };
         }
     };
+    let config_path: Option<String> = if settings.server_config.is_none() {
+        choose_config(server_kind)
+    } else {
+        settings.server_config.clone()
+    };
     let request_header: RequestHeader;
     let session_id: String;
     let user_id: String;
     let access_token: String;
+    let server_id: String;
     let mut device_id = uuid::Uuid::new_v4().to_string();
-    let mut config_file: ConfigFile;
-    let config_path = get_app_root(AppDataType::UserConfig, &APP_INFO).unwrap();
-    let config_path_string: String = if settings.server_config.is_none() {
-        format!("{}/{}.config.json", &config_path.display().to_string(), &media_server_name.to_lowercase())
-    } else {
-        settings.server_config.as_ref().unwrap().to_string()
-    }; 
-    if ! Path::exists(&config_path) || ! Path::new(&config_path_string).is_file() {
+    let config_file: ConfigFile;
+    if config_path.is_none() {
         app_root(AppDataType::UserConfig, &APP_INFO).expect("shit");
-        let ipaddress = configure_new_server();
+        let (ipaddress, server_name) = configure_new_server();
         let user_login = configure_new_login(media_server_name);
-        (auth_header, request_header, session_id, user_id, access_token) = test_auth(media_server_name, media_server, &ipaddress, &auth_header, &user_login, &device_id);
+        (auth_header, request_header, session_id, user_id, access_token, server_id) = test_auth(media_server_name, media_server, &ipaddress, &auth_header, &user_login, &device_id);
         config_file = ConfigFile { 
             emby,
             ipaddress,
@@ -151,42 +170,109 @@ pub fn check_information(settings: &Settings) -> HeadDict {
             access_token,
             username: user_login.username
         };
-        write_config(config_path_string, &config_file);
+        let config_path_string = generate_config_path(server_kind, server_id, server_name);
+        write_config(config_path_string, &config_file, None);
     } else {
+        let config_path_string = config_path.unwrap();
         println!("{}", "Configuration files found!".to_string().green());
-        let config_file_raw = read_config(&config_path_string, false);
-        if config_file_raw.is_some() {
-            config_file = config_file_raw.unwrap();
-            let ipaddress = &config_file.ipaddress;
-            device_id = config_file.device_id.clone();
-            auth_header = AuthHeader {
-                authorization: format!("Emby UserId={}, Client=\"Emby Theater\", Device=\"{}\", DeviceId=\"{}\", Version=\"{}\", Token={}", &config_file.user_id, APPNAME, device_id, VERSION, &config_file.access_token)
-            };
-            let session_id_test: Option<String> = re_auth(media_server_name, media_server, ipaddress, &auth_header, &device_id);
-            if session_id_test.is_none() {
-                println!("\nYour {} session expired! Please re-login.", media_server_name.to_lowercase());
+        let config_file_raw = read_config(&config_path_string, settings.autologin);
+        match config_file_raw {
+            Ok((mut file, mut raw_file)) => {
+                let ipaddress = &file.ipaddress;
+                device_id = file.device_id.clone();
+                auth_header = AuthHeader {
+                    authorization: format!("Emby UserId={}, Client=\"Emby Theater\", Device=\"{}\", DeviceId=\"{}\", Version=\"{}\", Token={}", &file.user_id, APPNAME, device_id, VERSION, &file.access_token)
+                };
+                let session_id_test: Option<String> = re_auth(media_server_name, media_server, ipaddress, &auth_header, &device_id);
+                if session_id_test.is_none() {
+                    println!("\nYour {} session expired! Please re-login.", media_server_name);
+                    let user_login = configure_new_login(media_server_name);
+                    (auth_header, request_header, session_id, _, access_token, _) = test_auth(media_server_name, media_server, ipaddress, &auth_header, &user_login, &device_id);
+                    file.access_token = access_token;
+                    let mut i = 0;
+                    while i < raw_file.user.len() {
+                        if raw_file.user[i].username == file.username {
+                            println!("Replaced {} in the config file.", raw_file.user.remove(i).username.green());
+                        } else {
+                            i += 1;
+                        }
+                    }
+                    write_config(config_path_string, &file, Some(raw_file.user));
+                } else {
+                    if raw_file.user[0].username != file.username {
+                        let mut i = 0;
+                        while i < raw_file.user.len() {
+                            if raw_file.user[i].username == file.username {
+                                println!("Set {} as the default user.", raw_file.user.remove(i).username.green());
+                            } else {
+                                i += 1;
+                            }
+                        };
+                        write_config(config_path_string, &file, Some(raw_file.user));
+                    };
+                    request_header = get_request_header(&file.access_token);
+                    session_id = session_id_test.unwrap();
+                }
+                config_file = ConfigFile {
+                    emby,
+                    device_id,
+                    ipaddress: ipaddress.to_string(),
+                    user_id: file.user_id,
+                    access_token: file.access_token,
+                    username: file.username
+                };
+            },
+            Err((Some(mut file), "add user")) => {
+                let ipaddress = file.ipaddress;
                 let user_login = configure_new_login(media_server_name);
-                (auth_header, request_header, session_id, _, access_token) = test_auth(media_server_name, media_server, ipaddress, &auth_header, &user_login, &device_id);
-                config_file.access_token = access_token;
-                write_config(config_path_string, &config_file);
-            } else {
-                request_header = get_request_header(&config_file.access_token);
-                session_id = session_id_test.unwrap();
+                (auth_header, request_header, session_id, user_id, access_token, _) = test_auth(media_server_name, media_server, &ipaddress, &auth_header, &user_login, &device_id);
+                config_file = ConfigFile {
+                    emby,
+                    device_id,
+                    ipaddress,
+                    user_id,
+                    access_token,
+                    username: user_login.username
+                };
+                let mut i = 0;
+                while i < file.user.len() {
+                    if file.user[i].username == config_file.username {
+                        println!("Replaced {}.", file.user.remove(i).username.green());
+                    } else {
+                        i += 1;
+                    }
+                };
+                write_config(config_path_string, &config_file, Some(file.user));
+            },
+            Err((None, "add server")) => {
+                let (ipaddress, _) = configure_new_server();
+                let user_login = configure_new_login(media_server_name);
+                (auth_header, request_header, session_id, user_id, access_token, _) = test_auth(media_server_name, media_server, &ipaddress, &auth_header, &user_login, &device_id);
+                config_file = ConfigFile {
+                    emby,
+                    device_id,
+                    ipaddress,
+                    user_id,
+                    access_token,
+                    username: user_login.username
+                };
+                write_config(config_path_string, &config_file, None);
+            },
+            _ => {
+                let (ipaddress, _) = configure_new_server();
+                let user_login = configure_new_login(media_server_name);
+                device_id = uuid::Uuid::new_v4().to_string();
+                (auth_header, request_header, session_id, user_id, access_token, _) = test_auth(media_server_name, media_server, &ipaddress, &auth_header, &user_login, &device_id);
+                config_file = ConfigFile {
+                    emby,
+                    device_id,
+                    ipaddress,
+                    user_id,
+                    access_token,
+                    username: user_login.username
+                };
+                write_config(config_path_string, &config_file, None);
             }
-        } else {
-            let ipaddress = configure_new_server();
-            let user_login = configure_new_login(media_server_name);
-            device_id = uuid::Uuid::new_v4().to_string();
-            (auth_header, request_header, session_id, user_id, access_token) = test_auth(media_server_name, media_server, &ipaddress, &auth_header, &user_login, &device_id);
-            config_file = ConfigFile {
-                emby,
-                device_id,
-                ipaddress,
-                user_id,
-                access_token,
-                username: user_login.username
-            };
-            write_config(config_path_string, &config_file);
         }
     }
     HeadDict {
@@ -200,8 +286,9 @@ pub fn check_information(settings: &Settings) -> HeadDict {
 }
 
 
-fn configure_new_server() -> String {
+fn configure_new_server() -> (String, String) {
     let mut ipaddress: String;
+    let mut server_name: String;
     println!("Searching for local media-server...");
     let socket:UdpSocket = UdpSocket::bind("0.0.0.0:0").expect("Failed to bind to network socket.");
     socket.set_read_timeout(Some(Duration::new(5, 0))).expect("nothing");
@@ -213,11 +300,13 @@ fn configure_new_server() -> String {
         Ok(_t) => {
             let parsed: UDP = byte_array_to_json(buf);
             ipaddress = parsed.Address;
-            print!("Is the media_server at the following address the correct one?\n \"{}\"\n (Y)es / (N)o", ipaddress);
+            server_name = parsed.Name;
+            print!("Is the {} at the following address the correct one?\n \"{}\"\n (Y)es / (N)o", server_name.green(), ipaddress);
             let udp_question = getch("YyNn");
             match udp_question {
                 'Y'|'y' => {
                     println!("Nice, already done.");
+
                 },
                 'N'|'n' => {
                     print!("Please specify the IP-Address manually\n(don't forget to add ports if not running on 80/443!)\n: ");
@@ -225,6 +314,11 @@ fn configure_new_server() -> String {
                     let mut ipaddress2 = String::new();
                     io::stdin().read_line(  &mut ipaddress2).unwrap();
                     ipaddress = ipaddress2.trim().parse().unwrap();
+                    print!("\nPlease enter a nickname for your media-server.\n(It's recommended to use a unique one)\n: ");
+                    io::stdout().flush().ok().expect("Failed to flush stdout");
+                    server_name = String::new();
+                    io::stdin().read_line(  &mut server_name).unwrap();
+                    server_name = server_name.trim().parse().unwrap();
                 }
                 _ => (),
             }
@@ -235,6 +329,11 @@ fn configure_new_server() -> String {
             let mut ipaddress2 = String::new();
             io::stdin().read_line(  &mut ipaddress2).unwrap();
             ipaddress = ipaddress2.trim().parse().unwrap();
+            print!("\nPlease enter a nickname for your media-server.\n(It's recommended to use a unique one)\n: ");
+            io::stdout().flush().ok().expect("Failed to flush stdout");
+            server_name = String::new();
+            io::stdin().read_line(  &mut server_name).unwrap();
+            server_name = server_name.trim().parse().unwrap();
         },
     }
     if ! ipaddress.contains("http") {
@@ -243,7 +342,7 @@ fn configure_new_server() -> String {
     if ipaddress.ends_with('/') {
         ipaddress.pop();
     }
-    ipaddress
+    (ipaddress, server_name)
 }
 
 
@@ -264,14 +363,13 @@ fn byte_array_to_json(buf: [u8; 4096]) -> UDP {
 fn configure_new_login(media_server_name: &str) -> UserLogin {
     fn take_input(media_server_name: &str) -> (String, String) {
         let mut username = String::new();
-        let mut password = String::new();
         print!("Please enter your {} username: ", media_server_name);
         io::stdout().flush().ok().expect("Failed to flush stdout");
         io::stdin().read_line(  &mut username).unwrap();
-        print!("Please enter your {} password: ", media_server_name);
+        print!("Please enter your {} password (hidden): ", media_server_name);
         io::stdout().flush().ok().expect("Failed to flush stdout");
-        io::stdin().read_line(  &mut password).unwrap();
-        println!();
+        let password = read_password().unwrap();
+        println!("");
         (password.trim().parse().unwrap(), username.trim().parse().unwrap())
     }
     let mut repeat: bool = true;
@@ -294,7 +392,7 @@ fn configure_new_login(media_server_name: &str) -> UserLogin {
 }
 
 
-fn test_auth (media_server_name: &str, media_server: &str, ipaddress: &String, auth_header: &AuthHeader, user_login: &UserLogin, device_id: &String) -> (AuthHeader, RequestHeader, String, String, String) {
+fn test_auth (media_server_name: &str, media_server: &str, ipaddress: &String, auth_header: &AuthHeader, user_login: &UserLogin, device_id: &String) -> (AuthHeader, RequestHeader, String, String, String, String) {
     println!("Testing {} connection ...", media_server_name);
     let username: String = user_login.username.clone();
     let password: String = user_login.pw.clone();
@@ -304,9 +402,10 @@ fn test_auth (media_server_name: &str, media_server: &str, ipaddress: &String, a
     match json_response {
         Ok(mut t) => {
             let json_response = t.json::<Value>().unwrap();
+            let server_id = json_response.get("ServerId").unwrap();
             let session_obj = json_response.get("SessionInfo").unwrap();
             let user_id = session_obj["UserId"].as_str().unwrap();
-            let session_id = &session_obj["Id"].as_str().unwrap();
+            let session_id = session_obj["Id"].as_str().unwrap();
             let token = json_response["AccessToken"].as_str().unwrap();
             (
                 AuthHeader {
@@ -319,6 +418,7 @@ fn test_auth (media_server_name: &str, media_server: &str, ipaddress: &String, a
                 session_id.to_string(),
                 user_id.to_string(),
                 token.to_string(),
+                server_id.as_str().unwrap().trim().to_string()
         )
         },
         _ => panic!("FCK")
@@ -342,48 +442,6 @@ pub fn post_puddler (url: String, auth_header: &AuthHeader, bod: String) -> Resu
         _ => panic!("{} fdsfds", response.status())
     };
     Ok(result)
-}
-
-
-fn read_config(config_path_string: &String, defaulted: bool) -> Option<ConfigFile> {
-    let file = std::fs::read_to_string(config_path_string).unwrap();
-    let local_config_file: Result<ConfigFile, serde_json::Error> = serde_json::from_str::<ConfigFile>(&file);
-    match local_config_file {
-        Ok(a) => {
-            let media_server_name: &str = if a.emby {
-                "Emby"
-            } else {
-                "Jellyfin"
-            };
-            if ! defaulted {
-                print!("Do you want to use this config?\n   Host ({}): {}\n   Username: {}\n (Y)es / (N)o", media_server_name, a.ipaddress, a.username);
-                let zhrtea = getch("YyNn");
-                io::stdout().flush().ok().expect("Failed to flush stdout");
-                match zhrtea {
-                    'Y' | 'y' => {
-                        Some(a)
-                    }
-                    'N' | 'n' => None,
-                    _ => None
-                }
-            } else {
-                Some(a)
-            }
-        },
-        Err(_) => {
-            println!("Config seems to be faulty.");
-            None
-        }
-    }
-}
-
-
-fn write_config(config_path_string: String, config_file: &ConfigFile) {
-    let result = std::fs::write(config_path_string, serde_json::to_string_pretty(&config_file).unwrap());
-    match result {
-        Ok(()) => println!("Saved to config file ..."),
-        Err(_e) => panic!("write access??")
-    }
 }