+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 ___
 
+## **0.5.6**
+
++ finally, full multi-server + multi-user support (config files have to be recreated)
++ added autologin feature since why not
++ updated all dependencies
++ move a little bit of code to the new `config.rs` where it belongs
++ password input will not be shown in the command line anymore
+
+___
+
 ## Note:
 + Below is to see what changed since I abandoned the python version.
 + bye